requirements.txt
click==8.0.3
colorama==0.4.4
Flask==2.0.2
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.2
joblib==1.1.0
MarkupSafe==2.0.1
numpy==1.21.3
pandas==1.3.4
patsy==0.5.2
pyserial==3.5
python-dateutil==2.8.2
pytz==2021.3
scikit-learn==1.0.1
scipy==1.7.1
six==1.16.0
sklearn==0.0
statsmodels==0.13.0
threadpoolctl==3.0.0
Werkzeug==2.0.2
